updated carousel to display 3 services
