+++ package.json
@@ -16,11 +16,13 @@
     "react": "^17.0.2",
     "react-bootstrap": "^2.0.2",
     "react-dom": "^17.0.2",
+    "react-elastic-carousel": "^0.11.5",
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.0.2",
     "react-scripts": "^0.9.5",
     "redux": "^4.1.2",
     "redux-thunk": "^2.4.0",
+    "styled-components": "^5.3.3",
     "use-validate-image-url": "^1.0.0",
     "web-vitals": "^1.0.1",
     "webpack": "^5.64.2",
