--- src/components/Carousel.js
@@ -1,24 +1,0 @@
-import React from 'react';
-import ServiceCards from './ServiceCards';
-
-function Carousel() {
-  return (
-    <div className="container w-50 mh-50">
-      <div id="carouselExampleCaptions" className="carousel slide" data-bs-ride="carousel">
-        <div className="carousel-inner ">
-          <ServiceCards />
-        </div>
-        <button className="carousel-control-prev" type="button" data-bs-target="#carouselExampleCaptions" data-bs-slide="prev">
-          <span className="carousel-control-prev-icon green" aria-hidden="true" />
-          <span className="visually-hidden">Previous</span>
-        </button>
-        <button className="carousel-control-next" type="button" data-bs-target="#carouselExampleCaptions" data-bs-slide="next">
-          <span className="carousel-control-next-icon green" aria-hidden="true" />
-          <span className="visually-hidden">Next</span>
-        </button>
-      </div>
-    </div>
-  );
-}
-
-export default Carousel;
