+++ src/index.css
@@ -97,3 +97,39 @@ i:hover {
 a:link {
   text-decoration: none;
 }
+
+.carouselcont {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.controls-wrapper {
+  width: 100%;
+}
+
+.carousel-wrapper {
+  width: 100%;
+}
+
+.cardcont {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+button.rec-dot{
+  background-color: #97c036;
+  box-shadow: 0 0 1px 3px #97c036;
+}
+
+button.rec-dot:hover, button.rec-dot:active, button.rec-dot:focus  {
+  box-shadow: 0 0 1px 3px #97c036;
+}
+
+button.rec.rec-arrow:hover,
+button.rec.rec-arrow:focus {
+  background-color: #97c036;
+}
+
